lab6/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
djangorestframework==3.12.2
drf-yasg==1.20.0
httpie==2.3.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.7
Pygments==2.7.3
pyparsing==2.4.7
PySocks==1.7.1
pytz==2020.4
requests==2.25.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.12
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
